Changed method maneuvers from int to size_t to appease -Wsign-compare

+++ aux3.c
@@ -1165,7 +1165,7 @@ void destroy_order()
 
 
 
-int maneuvers()
+size_t maneuvers()
 {
   int m;
